implemented CATR for bangla dataset

+++ engine.py
@@ -23,7 +23,7 @@ def train_one_epoch(model, criterion, data_loader,
             cap_masks = cap_masks.to(device)
 
             outputs = model(samples, caps[:, :-1], cap_masks[:, :-1])
-            loss = criterion(outputs.permute(0, 2, 1), caps[:, 1:])
+            loss = criterion(outputs.permute(0, 2, 1), caps[:, 1:].type(torch.int64))
             loss_value = loss.item()
             epoch_loss += loss_value
 
@@ -56,7 +56,7 @@ def evaluate(model, criterion, data_loader, device):
             cap_masks = cap_masks.to(device)
 
             outputs = model(samples, caps[:, :-1], cap_masks[:, :-1])
-            loss = criterion(outputs.permute(0, 2, 1), caps[:, 1:])
+            loss = criterion(outputs.permute(0, 2, 1), caps[:, 1:].type(torch.int64))
 
             validation_loss += loss.item()
